Add support for checking services by IP address
When given an IP address instead of a name, tls-checker attempted to
resolve the IP addresses corresponding to the provided value, which
produced no result.  As a consequence, testing a service by IP address
was not possible.

If an hostname can be cast as an IP address, consider it is one, and
return a single CertificateChecker for this address.  If the cast fails,
fall back to the original code which resolves the name into IP addresses
and return as many CertificateChecker as IP found.

This allows us to check unresolvable domains (e.g. because of DNSSEC):

       % bundle exec exe/tls-checker sigfail.verteiltesysteme.net
       % bundle exec exe/tls-checker 134.91.78.139
       {"state":"ok", ...}
       %

sigfail.verteiltesysteme.net has a deliberate invalid RRSIG to alow
checking DNS resolvers.  Check https://dnssec.vs.uni-due.de/ for more
details.