Fix depenency problems reported by CI
Remove unused include ActionView::Helpers::DateHelper.  Previous version
of the internet_security_event gem included actionview so this did not
raised an error when the code was moved from tls-checker to
internet_security_event.  However, the dependency was stripped down from
internet_security_event, so the inclusion is now failing.

The actionview gem in turn required logger, which is not automatically
loaded anymore, so load it before usage in the regression test suite.