Send critical events when TLS connection is impossible
We changed this behavior in 1.2.0, but it does not make sense to ignore
completely services for which a status is asked for (being passed as an
argument).

Networking error should better be handled at the Riemann level, by
passing stable events for 2 or 3 tests, and detecting states changes at
this point.