Fix support for Ruby 3.2+
Ruby 3.2 improved support of timeouts and when a timeout occur, a new
exception previously ignored is raised.

Rescue from such exceptions to preserve previous behavior.

While here, add the actual exception name in the message as it is
sometimes not sufficient, e.g. a Errno::ETIMEDOUT exception used to have
the message "Connection timed out - connect(2) for \"xxx\" port 443",
but a IO::TimeoutError exception have the message "connect(2) for
\"xxx\" port 443" without the root cause.